# UI/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mayra & Carlos API",
    "version": "v1"
  },
  "paths": {
    "/api/ConfirmAssitance/ConfirmAssistanceAsync": {
      "post": {
        "tags": [
          "ConfirmAssitance"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAssitanceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAssitanceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAssitanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ConfirmAssitance/ResendEmailAsync": {
      "post": {
        "tags": [
          "ConfirmAssitance"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAssitanceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAssitanceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAssitanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Family/GetAllByEmailAddresAsync/{emailAddres}": {
      "get": {
        "tags": [
          "Family"
        ],
        "parameters": [
          {
            "name": "emailAddres",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Family/GetAllByInvitationCodeAsync/{invitationCode}": {
      "get": {
        "tags": [
          "Family"
        ],
        "parameters": [
          {
            "name": "invitationCode",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Family/CreateFamilyAsync": {
      "post": {
        "tags": [
          "Family"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewFamilyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewFamilyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewFamilyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Family/UpdateFamily": {
      "post": {
        "tags": [
          "Family"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Family/GetOneByInvitationCodeAsync/{invitationCode}": {
      "get": {
        "tags": [
          "Family"
        ],
        "parameters": [
          {
            "name": "invitationCode",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/FamilyMember/AddFamilyMemberAsync": {
      "put": {
        "tags": [
          "FamilyMember"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMemberDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMemberDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FamilyMember/DeleteFamilyMember": {
      "delete": {
        "tags": [
          "FamilyMember"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMemberDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMemberDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfirmAssitanceDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "invitationCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "invitationCode": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "emailAddress": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "confirmationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "familyMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyMemberDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "familyId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "names": {
            "type": "string",
            "nullable": true
          },
          "isKid": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "NewFamilyDto": {
        "type": "object",
        "properties": {
          "invitationCode": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "emailAddress": {
            "type": "string",
            "nullable": true
          },
          "confirmationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "familyMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewFamilyMemberDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewFamilyMemberDto": {
        "type": "object",
        "properties": {
          "names": {
            "type": "string",
            "nullable": true
          },
          "isKid": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}